+++ HISTORY.rst
@@ -27,6 +27,7 @@ updated extract-dss command to used end of time period when resampling (min,max,
 Reads hydro h5 flows and area
 Godin filter time series
 command line tool :
+
  extract_dss to postprocess to daily, max, min and monthly from DSM2 dss files
 
 0.1.4 (2019-09-16)
@@ -34,6 +35,7 @@ command line tool :
 Reads hydro h5 flows and area
 Godin filter time series
 command line tool :
+
  extract_dss to postprocess to daily, max, min and monthly from DSM2 dss files
 
 0.1.0 (2019-08-21)